Refactor Junimo media imports and update component paths
Moved Junimo image files into a dedicated Junimos directory and created an index.ts for easier imports. Updated Achievements and Skills component paths to use index.ts files. Refactored stats.tsx to use new Junimo imports and updated import paths to use aliases for better maintainability.
