+++ src/Components/stats/stats.tsx
@@ -1,32 +1,40 @@
 import React, { useState } from 'react';
 import { Tab, Tabs, TabList, TabPanel } from 'react-tabs';
-import Skills from '../Skills/Skills'
-import Achievements from '../Achievements/Achievements' 
-import Junimo1a from '../../Media/Junimo1.png'
-import Junimo2a from '../../Media/Junimo2.png'
-import Junimo3a from '../../Media/Junimo3.png'
-import Junimo4a from '../../Media/Junimo4.png'
-import Junimo1  from '../../Media/Junimo1B.png'
-import Junimo2  from '../../Media/Junimo2B.png'
-import Junimo3  from '../../Media/Junimo3B.png' 
-import Junimo4  from '../../Media/Junimo4B.png' 
-import Rope     from '../../Media/Rope.png'
-import type { formattedSaveFileType, fullPlayerDataType } from 'types/displayDataTypes';
+import Skills from '@components/Skills';
+import Achievements from '@components/Achievements';
+import {
+    Junimo1,
+    Junimo2,
+    Junimo3,
+    Junimo4,
+    Junimo5 ,
+    Junimo6 ,
+    Junimo7 ,
+    Junimo8
+} from '@media/Junimos'
+import Rope from '@media/Rope.png'
+import type { formattedSaveFileType } from 'types/displayDataTypes';
 
 
 
 const Stats = ({playerData, farmName, farmhandData}: formattedSaveFileType) => {
     const [globalFarmName] = useState(farmName || "My Farm");
-    const [junimos] = useState([Junimo1, Junimo2, Junimo3, Junimo4]);
-    const [Activejunimos] = useState([Junimo1a, Junimo2a, Junimo3a, Junimo4a]);
+    const [junimos] = useState([
+    Junimo1,
+    Junimo2,
+    Junimo3,
+    Junimo4,
+    Junimo5 ,
+    Junimo6 ,
+    Junimo7 ,
+    Junimo8]);
 
     const titles = farmhandData.map((item, i) => 
         <Tab key={i}>
-            <img src={junimos[i + 1]} alt="Jun"></img> {item.playerName}
+            <img src={junimos[i + 1]} alt="player nametag"></img> {item.playerName}
         </Tab>
     );
     
-
     const playerStats = farmhandData.map((p, i) => (
         <TabPanel key={i}>
             <section className="wrapper">
