+++ src/Media/Junimos/index.ts
@@ -1,0 +1,9 @@
+
+export { default as Junimo1 } from './Junimo1.png'
+export { default as Junimo2 } from './Junimo2.png'
+export { default as Junimo3 } from './Junimo3.png'
+export { default as Junimo4 } from './Junimo4.png'
+export { default as Junimo5 }  from './Junimo1B.png'
+export { default as Junimo6 }  from './Junimo2B.png'
+export { default as Junimo7 }  from './Junimo3B.png' 
+export { default as Junimo8 }  from './Junimo4B.png'
